pyproject.toml
[project]
name = "koggi"
version = "0.0.3"
description = "PostgreSQL backup & restore CLI"
authors = [{name = "bamjun", email = "khy5116@naver.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["postgresql", "backup", "restore", "cli"]
dependencies = [
  "typer>=0.9.0",
  "rich>=13.0.0",
  "pydantic>=2.0.0",
  "psycopg2-binary>=2.9.0",
  "python-dotenv>=1.0.0"
]

[project.scripts]
koggi = "koggi.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
github = "https://github.com/bamjun/koggi"
